Fix team subscription test
